requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==5.0.7
django-import-export==4.1.1
idna==3.7
joblib==1.4.2
numpy==2.1.0
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
